Release 4.0.0 (#1492)
* Add UUIDModule, JavaDateTimeModule and RecordModule to support Java 17.
Since UUID fields, Date/DateTime fields are final baker cannot instantiate them in the pojo way.

* Remove java.util.Date

* Disabled the scala 2.12 compilation.

* Disabled the AkkaCassandraJmxMetricsSpec since it does not work with Java 17.

* Enabled the getIngredient endpoint for Bakery

* Made the java.util.Duration backwards compatible for Java versions.

* Disabled incompatible modules/plugins for Scala 2.13.

* Added support for the OffSetDateTime, ZonedDateTime & java.util.Date

* Set target back to Java 11 to ensure it does not break for java 11.

* Added missing function to Kotlin interface.

* Improve Kotlin DSL

* Added the getJavaBaker method to the kotlinDSL to ensure it can be used for frameworks that need a normal Baker.

* Added an optional rememberProcessDuration that forgets a process after the configured duration. This keeps the ProcessIndex shards from growing indefinitely.

* Fix to the ProcessDeleted not being handled during Bake.

* Made the lastupdate be set correctly after the RecipeInstance was used. (#1536)

* Reprovider functionality

* Fixed the cleanup for non Cassandra datastores. We use the original cleanup of the data as before our upgrade to the Cassandra cleanup tool for other datastores. (#1551)

* Wrote a type modules for the java.util.Currency.

* Added a section on Recipe Retention period to the docs (#1575)

* Remove generics for Place and Transition from PetriNet and IL (#1582)

* Fixed output ingredient rename example/doc for Java (#1574)

* Fixed output ingredient rename example/doc for Java

* fixed forgotten import

* Added some missing imports (#1583)

* Fixed the Wait interaction to work if you also have a retry strategy that fires and event. (#1585)

* Fixed a bug in the getOutputEventName where it would not work if the FailureStrategy would also throw an event.

* Added RecipeId and RecipeName to log lines (#1589)

* Set the default retention period to 14 days for deletion if the recipe is not found.

* Changed the behaviour of the FireEvent retry strategy. It now blocks the interaction instead of resolving it. This allows the interaction to be retried with the retyInteraction method.

* RecipeInstanceMetadata and RecipeInstanceEventList as ingredient input for interactions (#1643)

* Added extra logging if a recipe is not found during deletion of a process.

* Set the default retention period to 14 days for deletion if the recipe is not found.

* Changed the behaviour of the FireEvent retry strategy. It now blocks the interaction instead of resolving it. This allows the interaction to be retried with the retyInteraction method.

* Added the RecipeInstanceMetadata as input ingredient.

* Added a new Baker Event to represent internal interaction output events, checkpoint events and delayed transition events.

* Added the RecipeInstanceEventList as a input ingredient.

* Made the RecipeInstanceMetaData an internal concept instead of putting it in the ingredients and needing to transform it every time.

* Added validations that the Baker internal ingredients are not provided in the recipe.

* Filter out the CheckPoint event interactions from the visualisation

---------

+++ core/akka-runtime/src/main/scala/com/ing/baker/runtime/akka/actor/process_index/ProcessIndex.scala
@@ -13,6 +13,7 @@ import com.ing.baker.il.petrinet.{InteractionTransition, Place, Transition}
 import com.ing.baker.il.{CompiledRecipe, EventDescriptor}
 import com.ing.baker.petrinet.api._
 import com.ing.baker.runtime.akka._
+import com.ing.baker.runtime.akka.actor.{ActorBasedBakerCleanup, BakerCleanup, CassandraBakerCleanup}
 import com.ing.baker.runtime.akka.actor.Util.logging._
 import com.ing.baker.runtime.akka.actor.delayed_transition_actor.DelayedTransitionActor
 import com.ing.baker.runtime.akka.actor.delayed_transition_actor.DelayedTransitionActorProtocol.{FireDelayedTransition, StartTimer}
@@ -34,6 +35,8 @@ import com.ing.baker.runtime.serialization.Encryption
 import com.ing.baker.types.Value
 import com.typesafe.config.Config
 
+import java.util.concurrent.TimeUnit
+import scala.Console.println
 import scala.collection.mutable
 import scala.concurrent.duration._
 import scala.concurrent.{Await, ExecutionContext, Future}
@@ -49,7 +52,8 @@ object ProcessIndex {
             recipeManager: RecipeManager,
             getIngredientsFilter: Seq[String],
             providedIngredientFilter: Seq[String],
-            blacklistedProcesses: Seq[String]): Props =
+            blacklistedProcesses: Seq[String],
+            rememberProcessDuration: Option[Duration]): Props =
     Props(new ProcessIndex(
       recipeInstanceIdleTimeout,
       retentionCheckInterval,
@@ -58,7 +62,8 @@ object ProcessIndex {
       recipeManager,
       getIngredientsFilter,
       providedIngredientFilter,
-      blacklistedProcesses))
+      blacklistedProcesses,
+      rememberProcessDuration))
 
   sealed trait ProcessStatus
 
@@ -114,7 +119,8 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
                    recipeManager: RecipeManager,
                    getIngredientsFilter: Seq[String],
                    providedIngredientFilter: Seq[String],
-                   blacklistedProcesses: Seq[String]) extends PersistentActor with PersistentActorMetrics {
+                   blacklistedProcesses: Seq[String],
+                   rememberProcessDuration: Option[Duration]) extends PersistentActor with PersistentActorMetrics {
 
   override val log: DiagnosticLoggingAdapter = Logging.getLogger(logSource = this)
 
@@ -132,6 +138,7 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
   private val config: Config = context.system.settings.config
 
   private val snapShotInterval: Int = config.getInt("baker.actor.snapshot-interval")
+  private val snapshotCount: Int = config.getInt("baker.actor.snapshot-count")
 
   private val processInquireTimeout: FiniteDuration = config.getDuration("baker.process-inquire-timeout").toScala
   private val updateCacheTimeout: FiniteDuration =  config.getDuration("baker.process-index-update-cache-timeout").toScala
@@ -143,10 +150,17 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
   private val index: mutable.Map[String, ActorMetadata] = mutable.Map[String, ActorMetadata]()
   private val recipeCache: mutable.Map[String, (CompiledRecipe, Long)] = mutable.Map[String, (CompiledRecipe, Long)]()
 
-  private val cleanup = new akka.persistence.cassandra.cleanup.Cleanup(context.system)
+  //TODO chose if to use the CassandraBakerCleanup or the ActorBasedBakerCleanup
+  private val cleanup: BakerCleanup = {
+    if(config.hasPath("akka.persistence.journal.plugin") &&
+      config.getString("akka.persistence.journal.plugin") == "akka.persistence.cassandra.journal")
+      new CassandraBakerCleanup(context.system)
+    else
+      new ActorBasedBakerCleanup()
+  }
 
   private val delayedTransitionActor: ActorRef = context.actorOf(
-    props = DelayedTransitionActor.props(this.self),
+    props = DelayedTransitionActor.props(this.self, cleanup, snapShotInterval, snapshotCount),
     name = s"${self.path.name}-timer")
 
   // if there is a retention check interval defined we schedule a recurring message
@@ -163,6 +177,8 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
     log.debug("Added recipe to recipe cache: {}, current size: {}", recipeId, recipeCache.size)
     result.map(r => r.recipe -> r.updated)
   }
+
+  def getRecipeIdFromActor(actorRef: ActorRef) : String = actorRef.path.name
 
   def getRecipeWithTimeStamp(recipeId: String): Option[(CompiledRecipe, Long)] =
     recipeCache.get(recipeId) match {
@@ -191,15 +207,15 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
 
   // creates a ProcessInstanceActor, does not do any validation
   def createProcessActor(recipeInstanceId: String, compiledRecipe: CompiledRecipe): ActorRef = {
-    val runtime: ProcessInstanceRuntime[Place, Transition, RecipeInstanceState, EventInstance] =
+    val runtime: ProcessInstanceRuntime[RecipeInstanceState, EventInstance] =
       new RecipeRuntime(compiledRecipe, interactionManager, context.system.eventStream)
 
     val processActorProps =
       BackoffSupervisor.props(
         BackoffOpts.onStop(
-          ProcessInstance.props[Place, Transition, RecipeInstanceState, EventInstance](
+          ProcessInstance.props[RecipeInstanceState, EventInstance](
             compiledRecipe.name,
-            compiledRecipe.petriNet,
+            compiledRecipe,
             runtime,
             ProcessInstance.Settings(
               executionContext = bakerExecutionContext,
@@ -222,37 +238,61 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
   }
 
   def shouldDelete(meta: ActorMetadata): Boolean = {
-    meta.processStatus != Deleted &&
-      getCompiledRecipe(meta.recipeId)
-        .flatMap(_.retentionPeriod)
-        .exists { p => meta.createdDateTime + p.toMillis < System.currentTimeMillis() }
+    if(meta.processStatus != Deleted)
+      getCompiledRecipe(meta.recipeId) match {
+        case Some(recipe) =>
+          recipe.retentionPeriod.exists { p => meta.createdDateTime + p.toMillis < System.currentTimeMillis() }
+        case None =>
+          log.error(s"Could not find recipe: ${meta.recipeId} during deletion for recipeInstanceId: ${meta.recipeInstanceId} using default 14 days")
+          meta.createdDateTime + (14 days).toMillis < System.currentTimeMillis()
+      }
+    else false
   }
 
   private def deleteProcess(meta: ActorMetadata): Unit = {
-    getProcessActor(meta.recipeInstanceId) match {
-      case Some(actorRef: ActorRef) =>
-        log.debug(s"Deleting ${meta.recipeInstanceId} via actor message")
-        actorRef ! Stop(delete = true)
-      case None =>
-        log.debug(s"Deleting ${meta.recipeInstanceId} via cleanup tool")
-        getCompiledRecipe(meta.recipeId) match {
-          case Some(compiledRecipe) =>
-            val persistenceId = ProcessInstance.recipeInstanceId2PersistenceId(compiledRecipe.name, meta.recipeInstanceId)
-            log.debug(s"Deleting with persistenceId: ${persistenceId}")
-            persistWithSnapshot(ActorDeleted(meta.recipeInstanceId)) { _ =>
-              //Using deleteAllEvents since we do not use Snapshots for ProcessInstances
-              cleanup.deleteAllEvents(persistenceId, neverUsePersistenceIdAgain = false)
-                .map(_ -> {
-                  log.processHistoryDeletionSuccessful(meta.recipeInstanceId, 0)
-                  index.update(meta.recipeInstanceId, meta.copy(processStatus = Deleted))
-                })
-            }
-          case None =>
-            log.debug(s"Recipe not found for ${meta.recipeInstanceId}, marking as deleted")
-            persistWithSnapshot(ActorDeleted(meta.recipeInstanceId)) { _ =>
-              index.update(meta.recipeInstanceId, meta.copy(processStatus = Deleted))
-            }
+    //The new way of cleaning ProcessInstances, this can only be done if the datastore is Cassandra
+    if(cleanup.supportsCleanupOfStoppedActors) {
+      getProcessActor(meta.recipeInstanceId) match {
+        case Some(actorRef: ActorRef) =>
+          log.debug(s"Deleting ${meta.recipeInstanceId} via actor message")
+          actorRef ! Stop(delete = true)
+        case None =>
+          log.debug(s"Deleting ${meta.recipeInstanceId} via cleanup tool")
+          getCompiledRecipe(meta.recipeId) match {
+            case Some(compiledRecipe) =>
+              val persistenceId = ProcessInstance.recipeInstanceId2PersistenceId(compiledRecipe.name, meta.recipeInstanceId)
+              log.debug(s"Deleting with persistenceId: ${persistenceId}")
+              persistWithSnapshot(ActorDeleted(meta.recipeInstanceId)) { _ =>
+                //Using deleteAllEvents since we do not use Snapshots for ProcessInstances
+                cleanup.deleteAllEvents(persistenceId, neverUsePersistenceIdAgain = false)
+                  .map(_ -> {
+                    log.processHistoryDeletionSuccessful(meta.recipeInstanceId, 0)
+                    index.update(meta.recipeInstanceId, meta.copy(processStatus = Deleted))
+                  })
+              }
+            case None =>
+              log.debug(s"Recipe not found for ${meta.recipeInstanceId}, marking as deleted")
+              persistWithSnapshot(ActorDeleted(meta.recipeInstanceId)) { _ =>
+                index.update(meta.recipeInstanceId, meta.copy(processStatus = Deleted))
+              }
+          }
+      }
+    }
+    // The old way to cleanup ProcessInstances, we will let Akka Persistence handle the cleanup.
+    // This will first start the ProcessInstance actor even if is passivated so will have bigger memory impact and more reads on the datastore.
+    else {
+      getOrCreateProcessActor(meta.recipeInstanceId).foreach(_ ! Stop(delete = true))
+    }
+  }
+
+  private def forgetProcesses(): Unit = {
+    rememberProcessDuration.map {
+      duration: Duration =>
+        val currentTime = System.currentTimeMillis()
+        val toBeForgotten = index.filter { case (id, metadata) =>
+          metadata.isDeleted && currentTime >= (metadata.createdDateTime + duration.toMillis)
         }
+        index --= toBeForgotten.keys
     }
   }
 
@@ -282,10 +322,13 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
     } yield (transition, jobId)
   }
 
-  override def receiveCommand: Receive = {
-    case SaveSnapshotSuccess(_) => log.debug("Snapshot saved")
+  override def receiveCommand: Receive =  {
+    case SaveSnapshotSuccess(metadata) =>
+      log.debug("Snapshot saved & cleaning old processes")
+      cleanup.deleteEventsAndSnapshotBeforeSnapshot(metadata.persistenceId, snapshotCount)
 
-    case SaveSnapshotFailure(_, _) => log.error("Saving snapshot failed")
+    case SaveSnapshotFailure(_, _) =>
+      log.error("Saving snapshot failed")
 
     case GetIndex =>
       sender() ! Index(index.values.filter(_.processStatus == Active).toSeq)
@@ -295,9 +338,10 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
       if (toBeDeleted.nonEmpty)
         log.debug(s"Deleting recipe instance: {}", toBeDeleted.mkString(","))
       toBeDeleted.foreach(meta => deleteProcess(meta))
+      forgetProcesses()
 
     case Terminated(actorRef) =>
-      val recipeInstanceId = actorRef.path.name
+      val recipeInstanceId = getRecipeIdFromActor(actorRef)
 
       val mdc = Map(
         "processId" -> recipeInstanceId,
@@ -335,7 +379,7 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
               persistWithSnapshot(ActorCreated(recipeId, recipeInstanceId, createdTime)) { _ =>
 
                 // after that we actually create the ProcessInstance actor
-                val processState = RecipeInstanceState(recipeId, recipeInstanceId, Map.empty[String, Value], List.empty)
+                val processState = RecipeInstanceState(recipeId, recipeInstanceId, Map.empty[String, Value], Map.empty[String, String], List.empty)
                 val initializeCmd = Initialize(compiledRecipe.initialMarking, processState)
 
                 //TODO ensure the initialiseCMD is accepted before we add it ot the index
@@ -361,7 +405,7 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
           //Temporary solution for the situation that the initializeCmd is not send in the original Bake
           getCompiledRecipe(recipeId) match {
             case Some(compiledRecipe) =>
-              val processState = RecipeInstanceState(recipeId, recipeInstanceId, Map.empty[String, Value], List.empty)
+              val processState = RecipeInstanceState(recipeId, recipeInstanceId, Map.empty[String, Value], Map.empty[String, String], List.empty)
               val initializeCmd = Initialize(compiledRecipe.initialMarking, processState)
               createProcessActor(recipeInstanceId, compiledRecipe) ! initializeCmd
               sender() ! ProcessAlreadyExists(recipeInstanceId)
@@ -373,7 +417,7 @@ class ProcessIndex(recipeInstanceIdleTimeout: Option[FiniteDuration],
           //Temporary solution for the situation that the initializeCmd is not send in the original Bake
           getCompiledRecipe(recipeId) match {
             case Some(compiledRecipe) =>
-              val processState = RecipeInstanceState(recipeId, recipeInstanceId, Map.empty[String, Value], List.empty)
+              val processState = RecipeInstanceState(recipeId, recipeInstanceId, Map.empty[String, Value], Map.empty[String, String], List.empty)
               val initializeCmd = Initialize(compiledRecipe.initialMarking, processState)
               actorRef ! initializeCmd
               sender() ! ProcessAlreadyExists(recipeInstanceId)
